Files are uploded on 2/5/2023
The website was build and uploaded, I struggled a lot to layout the website. using float properties, In a particular situation, I misspelled float to flex and set it to float right and left, but it did not did anything, and after a long search for bug where the problem is , but finally I found that it was a stupid mistake of misspelling. at another situation I tried to 
fit the image to be width of 100% but it does not worked. After a while of struggling and I found that it was because of misspelled class name.